Refactored scenario testing to be better
