+++ packages/syft/setup.cfg
@@ -122,6 +122,10 @@ test_plugins =
     pytest-sugar
     pytest-lazy-fixture
     pytest-rerunfailures
+    pytest-asyncio
+    pytest-timeout
+    anyio
+    unsync
     coverage
     faker
     distro
