--- tests/scenario/bigquery/events.json
@@ -1,8 +1,0 @@
-{
-    "events": {
-        "users_created": {
-            "name": "users_created",
-            "event_time": 1724311872.780498
-        }
-    }
-}
