+++ tox.ini
@@ -17,6 +17,7 @@ envlist =
     syft.publish
     syft.test.security
     syft.test.unit
+    syft.test.scenario
     syft.test.notebook
     single_container.launch
     single_container.destroy
@@ -262,6 +263,18 @@ setenv =
 commands =
     bash -c 'ulimit -n 4096 || true'
     pytest -n auto --dist loadgroup --durations=20 --disable-warnings
+
+[testenv:syft.test.scenario]
+description = Syft Scenario Tests
+deps =
+    -e{toxinidir}/packages/syft[dev,data_science]
+allowlist_externals =
+    bash
+    uv
+changedir = {toxinidir}/tests/scenario
+setenv =
+commands =
+    pytest -s --disable-warnings
 
 [testenv:syft.test.notebook]
 description = Syft Notebook Tests
